Refactoring to split into several files
