+++ include/stream/SplitStream.h
@@ -1,0 +1,20 @@
+#ifndef _SPLIT_STREAM_H_
+#define _SPLIT_STREAM_H_
+
+#include "StreamTypes.h"
+
+template <typename T>
+class SplitStream: public Stream<T> {
+    int (*split_function) (T);
+    int num_streams;
+    Stream<T>** streams;
+public:
+    SplitStream(int (*split_function)(T), int num_streams, Stream<T>** streams) 
+        : split_function(split_function), num_streams(num_streams), streams(streams) {};
+
+    void receive(T value) {
+        int split_stream_num = split_function(value);
+        streams[split_stream_num % num_streams]->receive(value);
+    }
+};
+#endif
