+++ include/stream/Subscriber.h
@@ -1,0 +1,10 @@
+#ifndef _SUBSCRIBER_H_
+#define _SUBSCRIBER_H_
+
+template <typename T>
+class Subscriber {
+
+public:
+    virtual void receive(T) {};
+};
+#endif
